Checked to see if files/directories existed
Added an "if statement" to see if user.js and chrome existed before removing them

+++ spikes/Browser Features/removeFirefoxRestrictions.sh
@@ -7,14 +7,14 @@ echo "Script to revert Mozilla Firefox browser to previous settings"
 PROFDIR=$(cat ~/.mozilla/firefox/profiles.ini | grep 'Path=' | sed s/^Path=//)
 cd ~/.mozilla/firefox/$PROFDIR
 
-rm user.js
+[-e user.js ] && rm user.js
 touch user.js
 echo "user_pref(\"browser.tabs.closeWindowWithLastTab\", true);" >> user.js
 echo "user_pref(\"full-screen-api.enabled\", true);" >> user.js
 echo "user_pref(\"extensions.pocket.enabled\", true);" >> user.js
 echo "user_pref(\"browser.tabs.remote.separateFileUriProcess\", true);" >> user.js
 
-rm -r chrome
+[ -d chrome ] && rm -r chrome
 mkdir chrome
 cd chrome
 touch userContent.css
